requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# API framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Text processing
nltk>=3.8.1

# Utilities
python-multipart>=0.0.6
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
httpx>=0.25.0

# Logging
loguru>=0.7.0
keras
tensorflow

streamlit>=1.30.0
plotly>=5.15.0